refactor(core): break manager struct into smaller ones (#8124)
