+++ core/tauri/src/window/mod.rs
@@ -21,7 +21,7 @@ use crate::{
     CallbackFn, Invoke, InvokeBody, InvokeError, InvokeMessage, InvokeResolver,
     OwnedInvokeResponder,
   },
-  manager::WindowManager,
+  manager::AppManager,
   runtime::{
     monitor::Monitor as RuntimeMonitor,
     webview::{WebviewAttributes, WindowBuilder as _},
@@ -139,7 +139,7 @@ impl Monitor {
 /// A builder for a webview window managed by Tauri.
 #[default_runtime(crate::Wry, wry)]
 pub struct WindowBuilder<'a, R: Runtime> {
-  manager: WindowManager<R>,
+  manager: Arc<AppManager<R>>,
   runtime: RuntimeOrDispatch<'a, R>,
   app_handle: AppHandle<R>,
   label: String,
@@ -218,7 +218,7 @@ impl<'a, R: Runtime> WindowBuilder<'a, R> {
     let runtime = manager.runtime();
     let app_handle = manager.app_handle().clone();
     Self {
-      manager: manager.manager().clone(),
+      manager: manager.manager_owned(),
       runtime,
       app_handle,
       label: label.into(),
@@ -259,7 +259,7 @@ impl<'a, R: Runtime> WindowBuilder<'a, R> {
   /// [the Webview2 issue]: https://github.com/tauri-apps/wry/issues/583
   pub fn from_config<M: Manager<R>>(manager: &'a M, config: WindowConfig) -> Self {
     let builder = Self {
-      manager: manager.manager().clone(),
+      manager: manager.manager_owned(),
       runtime: manager.runtime(),
       app_handle: manager.app_handle().clone(),
       label: config.label.clone(),
@@ -455,9 +455,10 @@ impl<'a, R: Runtime> WindowBuilder<'a, R> {
         }));
     }
 
-    let labels = self.manager.labels().into_iter().collect::<Vec<_>>();
+    let labels = self.manager.window.labels().into_iter().collect::<Vec<_>>();
     let pending = self
       .manager
+      .window
       .prepare_window(self.app_handle.clone(), pending, &labels)?;
 
     #[cfg(desktop)]
@@ -472,6 +473,7 @@ impl<'a, R: Runtime> WindowBuilder<'a, R> {
     #[cfg(desktop)]
     let handler = self
       .manager
+      .menu
       .prepare_window_menu_creation_handler(window_menu.as_ref());
     #[cfg(not(desktop))]
     #[allow(clippy::type_complexity)]
@@ -484,7 +486,7 @@ impl<'a, R: Runtime> WindowBuilder<'a, R> {
       RuntimeOrDispatch::Dispatch(dispatcher) => dispatcher.create_window(pending, handler),
     }
     .map(|window| {
-      self.manager.attach_window(
+      self.manager.window.attach_window(
         self.app_handle.clone(),
         window,
         #[cfg(desktop)]
@@ -500,9 +502,9 @@ impl<'a, R: Runtime> WindowBuilder<'a, R> {
     if let Some(effects) = window_effects {
       crate::vibrancy::set_window_effects(&window, Some(effects))?;
     }
-    self.manager.eval_script_all(format!(
+    self.manager.window.eval_script_all(format!(
       "window.__TAURI_INTERNALS__.metadata.windows = {window_labels_array}.map(function (label) {{ return {{ label: label }} }})",
-      window_labels_array = serde_json::to_string(&self.manager.labels())?,
+      window_labels_array = serde_json::to_string(&self.manager.window.labels())?,
     ))?;
 
     self.manager.emit_filter(
@@ -976,7 +978,7 @@ pub struct Window<R: Runtime> {
   /// The webview window created by the runtime.
   pub(crate) window: DetachedWindow<EventLoopMessage, R>,
   /// The manager to associate this webview window with.
-  pub(crate) manager: WindowManager<R>,
+  pub(crate) manager: Arc<AppManager<R>>,
   pub(crate) app_handle: AppHandle<R>,
   js_event_listeners: Arc<Mutex<HashMap<JsEventListenerKey, HashSet<EventId>>>>,
   // The menu set for this window
@@ -1057,8 +1059,12 @@ impl<R: Runtime> Manager<R> for Window<R> {
   }
 }
 impl<R: Runtime> ManagerBase<R> for Window<R> {
-  fn manager(&self) -> &WindowManager<R> {
+  fn manager(&self) -> &AppManager<R> {
     &self.manager
+  }
+
+  fn manager_owned(&self) -> Arc<AppManager<R>> {
+    self.manager.clone()
   }
 
   fn runtime(&self) -> RuntimeOrDispatch<'_, R> {
@@ -1162,7 +1168,7 @@ impl PlatformWebview {
 impl<R: Runtime> Window<R> {
   /// Create a new window that is attached to the manager.
   pub(crate) fn new(
-    manager: WindowManager<R>,
+    manager: Arc<AppManager<R>>,
     window: DetachedWindow<EventLoopMessage, R>,
     app_handle: AppHandle<R>,
     #[cfg(desktop)] menu: Option<WindowMenu<R>>,
@@ -1323,8 +1329,8 @@ impl<R: Runtime> Window<R> {
   ) {
     self
       .manager
-      .inner
-      .window_menu_event_listeners
+      .menu
+      .event_listeners
       .lock()
       .unwrap()
       .insert(self.label().to_string(), Box::new(f));
@@ -1367,7 +1373,7 @@ impl<R: Runtime> Window<R> {
   pub fn set_menu(&self, menu: Menu<R>) -> crate::Result<Option<Menu<R>>> {
     let prev_menu = self.remove_menu()?;
 
-    self.manager.insert_menu_into_stash(&menu);
+    self.manager.menu.insert_menu_into_stash(&menu);
 
     let window = self.clone();
     let menu_ = menu.clone();
@@ -2210,7 +2216,7 @@ impl<R: Runtime> Window<R> {
       }
     };
 
-    let custom_responder = self.manager.invoke_responder();
+    let custom_responder = self.manager.window.invoke_responder.clone();
 
     let resolver = InvokeResolver::new(
       self.clone(),
